+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Add a `python_requires` to enforce support of Python 3.6+ only.
+
 ## [v0.6.0](https://github.com/thibaudcolas/curlylint/releases/tag/v0.6.0) 2020-03-30
 
 ### Changed